ci: remove incorrect support for `inputs.name`
The `test.yaml` workflow was using `inputs.name`, but it only
worked because the `name` input was not defined in the
`starship-action`.

When `inputs.name` support was implemented in PR #42, it broke
the `test.yaml` workflow because the `starship stop --name`
setting overrode the `config.yml` `name:` value.  This
installed the Helm release with a different name than what was
being stopped.

Without stopping the correct Helm release, `starship stop`
would hang indefinitely.  Now it doesn't.